Changed the function to retrieve subtypes of AsyncBaseOp not be pure virtual
The function to retrieve supported subclasses was pure virtual. This is a bit cumbersome
when subtypes are implemented since they return a null for all the other subtypes except
for themselves. Instead now we have a default implementation which returns null while the
subtypes only override the function specific to their type.